redditdl: use the new saver
